Fix infinite loop on premature EOF when decompressing.
Clarify the comments about one assertion and what that tells us
about premature EOF conditions.
Unlike zlib, bzlib doesn't have a BufferError return value to indcate
when no forward progress was possible. So we can't use that to detect
EOF. So we have to do it manally by checking if the input buffer is 
empty and the output buffer has free space left in it and there are no
more input chunks left.